Rename fmriprep testing utils
